requirements-venv.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
anthropic==0.75.0
anyio==3.7.1
attrs==25.4.0
backoff==2.2.1
bcrypt==4.1.1
black==23.12.0
blinker==1.9.0
cachelib==0.13.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.0
cryptography==41.0.7
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.3.1
filelock==3.20.1
flake8==6.1.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-Session==0.5.0
frozenlist==1.8.0
fsspec==2025.12.0
google-api-core==2.28.1
google-api-python-client==2.104.0
google-auth==2.45.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mistralai==1.10.0
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.4.0
oauthlib==3.3.1
ollama==0.6.1
openai==2.14.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.1.0
pyotp==2.9.0
pyparsing==3.3.1
pypng==0.20220715.0
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
Werkzeug==3.1.4
yarl==1.22.0
zipp==3.23.0
